environment variable passed from install script to genconf script

+++ dreamland/genconf.sh
@@ -23,25 +23,26 @@ ln -sf ~/Projects/dotfiles/dreamland/vimperatorrc_conf ~/.vimperatorrc
 
 sudo ln -sf ~/Projects/dotfiles/dreamland/libinput_conf /etc/X11/xorg.conf.d/50-touchpad.conf
 
-sudo ln -sf ~/Projects/dotfiles/dreamland/virtualbox_conf /etc/modules-load.d/virtualbox.conf
 
 ln -sf ~/Projects/dotfiles/dreamland/asoundrc_conf ~/.asoundrc
 
 ln -sf ~/Projects/dotfiles/dreamland/tmux_conf ~/.tmux.conf
 
-sudo ln -sf ~/Projects/dotfiles/dreamland/bumblebee_nvidia_conf /etc/bumblebee/xorg.conf.nvidia
-
-sudo ln -sf ~/Projects/dotfiles/dreamland/tlp_conf /etc/default/tlp
-sudo systemctl enable tlp.service
-sudo systemctl enable tlp-sleep.service
-sudo systemctl disable systemd-rfkill.service
+if [ $environment == "real" ]
+    then
+        sudo ln -sf /home/anhtuann/Projects/dotfiles/dreamland/bumblebee_conf /etc/bumblebee/bumblebee.conf
+        sudo ln -sf ~/Projects/dotfiles/dreamland/bumblebee_nvidia_conf /etc/bumblebee/xorg.conf.nvidia
+        sudo ln -sf ~/Projects/dotfiles/dreamland/tlp_conf /etc/default/tlp
+        sudo systemctl enable tlp.service
+        sudo systemctl enable tlp-sleep.service
+        sudo systemctl disable systemd-rfkill.service
+        sudo ln -sf ~/Projects/dotfiles/dreamland/virtualbox_conf /etc/modules-load.d/virtualbox.conf
+fi
 
 sudo ln -sf ~/Projects/dotfiles/dreamland/pacman_conf /etc/pacman.conf
 
 mkdir ~/.config/dunst
 ln -sf ~/Projects/dotfiles/dreamland/dunst_conf /home/anhtuann/.config/dunst/dunstrc
-
-sudo ln -sf /home/anhtuann/Projects/dotfiles/dreamland/bumblebee_conf /etc/bumblebee/bumblebee.conf
 
 ranger --copy-config=all
 ln -sf /home/anhtuann/Projects/dotfiles/dreamland/ranger_rifle_conf /home/anhtuann/.config/ranger/rifle.conf
